requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.4
openpyxl==3.1.2
numpy==1.26.2
python-dateutil==2.8.2
streamlit==1.29.0
plotly==5.18.0
altair==5.2.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
python-dotenv==1.0.0
psutil==6.1.1
requests==2.32.3
urllib3==2.3.0
gspread==5.12.2
oauth2client==4.1.3
dash==2.14.2
google-auth-oauthlib==1.2.0
google-auth==2.25.2
watchdog==3.0.0
